+++ system/sso.php
@@ -1,0 +1,7 @@
+<?php
+return [
+  'enabled' => false,
+  // OIDC / SAML foundation configs are stored in DB; this file is for defaults and feature flags.
+  'oidc' => ['enabled' => true],
+  'saml' => ['enabled' => true],
+];